+++ protected/humhub/assets/JqueryNiceScrollAsset.php
@@ -8,24 +8,19 @@
 
 namespace humhub\assets;
 
-use humhub\components\assets\AssetBundle;
+use humhub\components\assets\WebStaticAssetBundle;
 
 /**
  * jQuery-nicescroll
  *
  * @author luke
  */
-class JqueryNiceScrollAsset extends AssetBundle
+class JqueryNiceScrollAsset extends WebStaticAssetBundle
 {
 
     /**
      * @inheritdoc
      */
-    public $sourcePath = '@npm/jquery.nicescroll';
-
-    /**
-     * @inheritdoc
-     */
-    public $js = ['jquery.nicescroll.min.js'];
+    public $js = ['js/jquery.nicescroll.min.js'];
 
 }