+++ protected/humhub/assets/FlatelementsStyleAsset.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace humhub\assets;
+
+use humhub\components\assets\WebStaticAssetBundle;
+
+class FlatelementsStyleAsset extends WebStaticAssetBundle
+{
+    public $css = [
+        'css/flatelements.css'
+    ];
+}