+++ protected/humhub/assets/HighlightJsStyleAsset.php
@@ -11,18 +11,11 @@ namespace humhub\assets;
 
 use humhub\components\assets\WebStaticAssetBundle;
 
-class HighlightJsAsset extends WebStaticAssetBundle
+class HighlightJsStyleAsset extends WebStaticAssetBundle
 {
 
     /**
      * @inheritdoc
      */
-    public $js = ['js/highlight.js/highlight.pack.js'];
-
-    /**
-     * @inheritdoc
-     */
-    public $depends = [
-        HighlightJsStyleAsset::class
-    ];
+    public $css = ['js/highlight.js/styles/github.css'];
 }