+++ protected/humhub/modules/content/models/Content.php
@@ -17,6 +17,7 @@ use humhub\modules\content\components\ContentActiveRecord;
 use humhub\modules\content\components\ContentContainerActiveRecord;
 use humhub\modules\content\components\ContentContainerModule;
 use humhub\modules\content\interfaces\ContentOwner;
+use humhub\modules\content\live\NewContent;
 use humhub\modules\content\permissions\CreatePrivateContent;
 use humhub\modules\content\permissions\CreatePublicContent;
 use humhub\modules\content\permissions\ManageContent;
@@ -198,7 +199,7 @@ class Content extends ActiveRecord implements Movable, ContentOwner
         }
 
         if ($this->container) {
-            Yii::$app->live->send(new \humhub\modules\content\live\NewContent([
+            Yii::$app->live->send(new NewContent([
                 'sguid' => ($this->container instanceof Space) ? $this->container->guid : null,
                 'uguid' => ($this->container instanceof User) ? $this->container->guid : null,
                 'originator' => $this->createdBy->guid,
@@ -208,7 +209,8 @@ class Content extends ActiveRecord implements Movable, ContentOwner
                 'sourceId' => $contentSource->getPrimaryKey(),
                 'silent' => $this->isMuted(),
                 'streamChannel' => $this->stream_channel,
-                'contentId' => $this->id
+                'contentId' => $this->id,
+                'insert' => $insert
             ]));
         }