+++ protected/humhub/docs/CHANGELOG.md
@@ -1,8 +1,10 @@
 HumHub Change Log
 =================
 
-1.5.0-beta.2 (Unreleased)
--------------------------
+
+1.5.0 (April 15, 2020)
+----------------------
+
 - Fix #3976: User profile stream could not be initialized
 - Fix #3977 Reverted nicescroll version to 3.6.6 due to mobile scroll issues
 - Chg: Only use nicescroll on space chooser, notifications, and activity stream on large devices
@@ -11,6 +13,7 @@ HumHub Change Log
 - Fix #3925: "Back to stream" button visible after pin to top in stream
 - Chg: Removed all style assets from CoreBundleAssets
 - Fix #3991: Archive content triggers "New updates available" stream badge
+
 
 1.5.0-beta.1 (April 6, 2020)
 ----------------------------